Move standard ctype sets to repn.util
