+++ pyomo/repn/util.py
